# backend/src/main/resources/static/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Miometry API
  description: |
    REST API for Miometry enabling engineers to record daily work hours,
    manage absences, submit monthly time entries for approval, and perform bulk operations.
    
    ## Authentication
    All endpoints require OAuth2/SAML2 authentication via organizational SSO provider.
    Bearer token must be included in Authorization header.
    
    ## Rate Limiting
    - Standard endpoints: 1000 requests/hour per user
    - CSV import: 10 concurrent uploads per user
    - Streaming endpoints: 100 concurrent connections per user
    
    ## Versioning
    API version is specified in URL path: /api/v1/*
  version: 1.0.0
  contact:
    name: Miometry Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.worklog.example.com/api/v1
    description: Production
security:
  - BearerAuth: []
paths:
...
  /worklog/csv/import:
    post:
      summary: Import work log entries from CSV
      description: |
        Streaming CSV import with progress feedback.
        Accepts CSV files up to 100,000 rows (≈10MB) via UI.
        Larger files should use streaming backend API (future).
        
        CSV Format:
        date,projectCode,hours,comment
        2026-01-15,PROJ-001,8.00,"Development work"
      tags:
        - CSV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file
                memberId:
                  type: string
                  format: uuid
                  description: Target member (for proxy entry)
      responses:
        '200':
          description: Import completed with results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsvImportResponse'
        '400':
          description: Invalid CSV format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: File too large (>10MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /worklog/csv/import/{importId}/progress:
    get:
      summary: Get CSV import progress (SSE)
      description: Server-Sent Events endpoint for real-time import progress
      tags:
        - CSV
      parameters:
        - name: importId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: SSE stream of progress updates
          content:
            text/event-stream:
              schema:
                type: object
                properties:
                  rowsProcessed:
                    type: integer
                  totalRows:
                    type: integer
                  percentage:
                    type: number
                    format: float
                  estimatedTimeRemaining:
                    type: string
                    format: duration
                  status:
                    type: string
                    enum: [PROCESSING, COMPLETED, FAILED]
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth2/SAML2 access token
...
  schemas:
...
    CsvImportResponse:
      type: object
      properties:
        importId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PROCESSING, COMPLETED, FAILED, PARTIAL]
        totalRows:
          type: integer
        successfulRows:
          type: integer
        failedRows:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              rowNumber:
                type: integer
              error:
                type: string
              data:
                type: string
        processingTimeMs:
          type: integer
          format: int64
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
...